requirements.txt
streamlit
Pillow
folium
streamlit-folium
sqlite3
